# services/http-api/api/static/docs/swagger.yaml
basePath: /v1
...
host: localhost:9001
info:
  contact: {}
  description: Bringing connectivity and automation to everyone, everywhere.
  title: IOTEA
  version: alpha-v1.0
paths:
...
  /channels/{channelId}:
    delete:
      consumes:
      - application/json
      description: Given a channel ID, delete a channel. Channel must be unpublished
        before it can be deleted.
      parameters:
      - description: Space ID
        in: query
        name: spaceId
        required: true
        type: string
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete a channel
      tags:
      - channels
    get:
      consumes:
      - application/json
      description: Given a channel ID, retrieve the channel's details.
      parameters:
      - description: Space ID
        in: query
        name: spaceId
        required: true
        type: string
      - description: Channel ID
        in: path
        name: channelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get a channel
      tags:
      - channels
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"